+++ pom.xml
@@ -126,19 +126,19 @@
 			<version>1.2.16</version>
 		</dependency>
 
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>5.0-SNAPSHOT</version>
-			<type>jar</type>
-		</dependency>
-
 		<!--<dependency>-->
 			<!--<groupId>junit</groupId>-->
 			<!--<artifactId>junit</artifactId>-->
-			<!--<version>4.12</version>-->
-			<!--<scope>test</scope>-->
+			<!--<version>5.0-SNAPSHOT</version>-->
+			<!--<type>jar</type>-->
 		<!--</dependency>-->
+
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>4.12</version>
+			<scope>test</scope>
+		</dependency>
 		
 		<dependency>
             <groupId>org.apache.camel</groupId>
@@ -225,38 +225,14 @@
 			<version>4.0.1-alpha</version>
 			<type>jar</type>
 		</dependency>
+
 		<dependency>
 			<groupId>org.testfx</groupId>
 			<artifactId>testfx-junit</artifactId>
 			<version>4.0.1-alpha</version>
 			<type>jar</type>
 		</dependency>
-		<dependency>
-			<groupId>com.google.guava</groupId>
-			<artifactId>guava</artifactId>
-			<version>18.0</version>
-		</dependency>
 
-		<!--<dependency>-->
-			<!--<groupId>org.jfxtras</groupId>-->
-			<!--<artifactId>openjfx-monocle</artifactId>-->
-			<!--<version>1.8.0_20</version>-->
-			<!--<scope>test</scope>-->
-		<!--</dependency>-->
-
-		<!--<dependency>-->
-			<!--<groupId>org.loadui</groupId>-->
-			<!--<artifactId>testFx</artifactId>-->
-			<!--<version>3.1.0</version>-->
-			<!--<scope>test</scope>-->
-		<!--</dependency>-->
-
-
-		<dependency>
-			<groupId>org.controlsfx</groupId>
-			<artifactId>controlsfx</artifactId>
-			<version>8.40.10</version>
-		</dependency>
 	</dependencies>
 
 	<profiles>
