+++ src/test/java/dpfmanager/gui/ImportConfigFileTest.java
@@ -1,0 +1,7 @@
+package dpfmanager.gui;
+
+/**
+ * Created by Adrià Llorens on 07/01/2016.
+ */
+public class ImportConfigFileTest {
+}
